% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook.R
\name{add_hook}
\alias{add_hook}
\title{Add a user hook to be executed before launching an R subprocess}
\usage{
add_hook(...)
}
\arguments{
\item{...}{Named argument specifying a hook function to add, or \code{NULL} to
delete the named hook.}
}
\value{
\code{add_hook} is called for its side-effects.
}
\description{
This function allows users of \code{callr} to specify functions that get invoked
whenever an R session is launched. The function can modify the environment
variables and command line arguments.
}
\details{
The prototype of the hook function is \verb{function (options)}, and it is
expected to return the modified \code{options}.
}
