% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callr-package.R
\docType{package}
\name{callr-package}
\alias{callr}
\alias{callr-package}
\title{Call R from R}
\description{
It is sometimes useful to perform a computation in a separate
R process, without affecting the current R process at all.  This
packages does exactly that.
}
\section{callr}{
\if{html}{\out{<div class="sourceCode {r man-readme, child = "README.Rmd"}">}}\preformatted{}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://callr.r-lib.org}
  \item \url{https://github.com/r-lib/callr}
  \item Report bugs at \url{https://github.com/r-lib/callr/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com} (\href{https://orcid.org/0000-0001-7098-9676}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Gábor Csárdi \email{csardi.gabor@gmail.com} (\href{https://orcid.org/0000-0001-7098-9676}{ORCID}) [copyright holder]
  \item Winston Chang
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item Ascent Digital Services [copyright holder, funder]
}

}
\keyword{internal}
