% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/callr.R
\name{callr}
\alias{callr}
\title{Call an R function via JSON}
\usage{
callr(filename_in, filename_out = NULL, backup = FALSE, prettify = TRUE,
  auto_unbox = TRUE)
}
\arguments{
\item{filename_in}{Filename for JSON with function to run}

\item{filename_out}{Filename to write the output to}

\item{backup}{Backup existing file when}

\item{prettify}{Produce pretty json output?}

\item{auto_unbox}{Turn length-1 vectors into scalars in json output?}
}
\description{
Call an R function via JSON
}

