% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{convert_and_check_my_args}
\alias{convert_and_check_my_args}
\title{Convert and check function arguments}
\usage{
convert_and_check_my_args(options)
}
\arguments{
\item{options}{List of options.}
}
\description{
This function is used for all variants of \code{r} and \code{rcmd}. An argument
name is only used to refer to one kind of object, to make this possible.
}
\details{
The benefit of having a single \code{options} object is to avoid passing
around a lot of arguments all the time.

The benefit of making this object internal (i.e. that the \code{r}, etc.
functions have multiple arguments instead of a single \code{options} list),
is that documentation and usage is more user friendly (e.g. command-
completion works in the editor.
}
\keyword{internal}
