% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_repos}
\alias{default_repos}
\title{Default value for the \code{repos} option in callr subprocesses}
\usage{
default_repos()
}
\value{
Named character vector, the default value of the \code{repos}
option in callr subprocesses.
}
\description{
callr sets the \code{repos} option in subprocesses, to make sure that
a CRAN mirror is set up. This is because the subprocess cannot bring
up the menu of CRAN mirrors for the user to choose from.
}
\examples{
default_repos()
}
