% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callr.R
\name{install}
\alias{install}
\title{Install callr script}
\usage{
install(path)
}
\arguments{
\item{path}{Directory to install the script}
}
\description{
Installs the "callr" script in the directory \code{path}
}

