% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install packages}
\usage{
install_packages(packages, package_sources = NULL, missing_only = TRUE)
}
\arguments{
\item{packages}{Vector of packages}

\item{package_sources}{List with details for installation of
non-CRAN sources.}

\item{missing_only}{Install only missing packages?}
}
\description{
Install packages, including from non-CRAN sources
}
\details{
\code{package_sources} must be a named list; the list corresponds
to package names.  Each element is a list, most likely with
elements \code{source="github"} and \code{repo="username/repo"}.
}

