% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{load_packages}
\alias{load_packages}
\title{Load a set of packages}
\usage{
load_packages(packages, package_sources = NULL, quiet = TRUE,
  install_missing = FALSE)
}
\arguments{
\item{packages}{Character vector with a set of packages to load}

\item{package_sources}{Information on where to install additional
package from (see \code{\link{install_packages}}}

\item{quiet}{Load packages quietly?}

\item{install_missing}{Attempt to install missing packages?}
}
\description{
Load a set of packages, possibly installing first.
}

