% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{load_source_files}
\alias{load_source_files}
\title{Load a set of source files}
\usage{
load_source_files(source_files, envir = .GlobalEnv, packages = character(0),
  ...)
}
\arguments{
\item{source_files}{Characte vector with a set of source files to
load.  Paths should be relative to the current working directory,
though absolute paths are allowed.}

\item{envir}{Environment to source into.  By default, sources into
the global environment (as does \code{\link{source}}, but in
contrast to \code{\link{sys.source}}).}

\item{packages}{Packages to load before sourcing the files.  This
is perhaps useful for code that uses the magrittr pipe to define
functions, or for source files that generate objects as opposed to
functions.}

\item{...}{Additional arguments to \code{\link{sys.source}} (i.e.,
\code{chdir} and \code{keep.source}).}
}
\description{
Load a set of source files.  The function \code{load_source_files}
is roughly equivalent to running a for loop over a set of source
files, except that it preloads some packages, keeps a hash of the
source files that were generated, has slightly better error
messaging and is focussed on creating environments.
}
\details{
Note that the if sourced files source other files then the
environment will not easily be recreatable (this is due to the
difficulty of redirecting where source is pointing to in
subsequent calls).
}

