% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{new_callr_error}
\alias{new_callr_error}
\title{Create an error object}
\usage{
new_callr_error(out, msg = NULL)
}
\arguments{
\item{out}{The object returned by \code{\link[=run]{run()}}.}

\item{msg}{An extra message to add to the error message.}
}
\description{
There are two kinds of errors, both have class \code{callr_error}:
\enumerate{
\item the first one is thrown after a timeout: \code{callr_timeout_error}.
\item the second one is thrown after an R error (in the other session):
\code{callr_status_error}.
}

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\keyword{internal}
