% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-initialize.R
\name{process_initialize}
\alias{process_initialize}
\title{Start a process}
\usage{
process_initialize(self, private, command, args, stdout, stderr, cleanup,
  echo_cmd, supervise, windows_verbatim_args, windows_hide_window, encoding)
}
\arguments{
\item{self}{this}

\item{private}{this$private}

\item{command}{Command to run, string scalar.}

\item{args}{Command arguments, character vector.}

\item{stdout}{Standard output, FALSE to ignore, TRUE for temp file.}

\item{stderr}{Standard error, FALSE to ignore, TRUE for temp file.}

\item{cleanup}{Kill on GC?}

\item{echo_cmd}{Echo command before starting it?}

\item{supervise}{Should the process be supervised?}

\item{encoding}{Assumed stdout and stderr encoding.}
}
\description{
Start a process
}
\keyword{internal}
