% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presets.R
\name{r_copycat}
\alias{r_copycat}
\title{Run an R process that mimics the current R process}
\usage{
r_copycat(func, args = list(), libpath = .libPaths(),
  repos = getOption("repos"), cmdargs = "--slave", system_profile = TRUE,
  user_profile = TRUE, env = character(), ...)
}
\description{
Differences to \code{r()}:
\itemize{
\item No extra repoditories are set up.
\item The \code{--no-site-file}, \code{--no-environ}, \code{--no-save}, \code{--no-restore}
command line arguments are not used. (But \code{--slave} still is.)
\item The system profile and the user profile are loaded.
\item No extra environment variables are set up.
}
}
\seealso{
Other callr functions: \code{\link{r_vanilla}},
  \code{\link{r}}
}
