% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{r_process_options}
\alias{r_process_options}
\title{Create options for an \link{r_process} object}
\usage{
r_process_options(...)
}
\arguments{
\item{...}{Options to override, named arguments.}
}
\value{
A list of options.

\code{r_process_options()} creates a set of options to initialize a new
object from the \code{r_process} class. Its arguments must be named, the
names are used as option names. The options correspond to (some of)
the arguments of the \code{\link[=r]{r()}} function. At least the \code{func} option must be
specified, this is the R function to run in the background.
}
\description{
Create options for an \link{r_process} object
}
\examples{
## List all options and their default values:
r_process_options()
}
