% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmd.R
\name{rcmd}
\alias{rcmd}
\title{Run an R CMD command}
\usage{
rcmd(cmd, cmdargs = character(), libpath = .libPaths(),
  repos = getOption("repos"), stdout = NULL, stderr = NULL)
}
\arguments{
\item{cmd}{Command to run. See \code{R --help} from the command
line for the various commands. In the current version of R (3.2.4)
these are: BATCH, COMPILE, SHLIB, INSTALL, REMOVE, build, check,
LINK, Rprof, Rdconv, Rd2pdf, Rd2txt, Stangle, Sweave, Rdiff, config,
javareconf, rtags.}

\item{cmdargs}{Command line arguments.}

\item{libpath}{R library path to set up using environment variables.
Defaults to the current library path.}

\item{repos}{The \code{repos} option to set CRAN mirrors and other
CRAN-like repositories.}

\item{stdout}{Optionally a file name to send the standard output to.}

\item{stderr}{Optionally a file name to send the standard error to.}
}
\value{
A list with the standard output (\code{$stdout}), standard
  error (\code{stderr}) and exit status (\code{$status}) of the
  external \code{R CMD} command.
}
\description{
Run an R CMD command form within R. This will usually start
another R process, from a shell script.
}
\examples{
rcmd("config", "CC")
}

