% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processx-forward.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{run}
\alias{process}
\alias{poll}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{processx}{\code{\link[processx]{poll}}, \code{\link[processx]{process}}, \code{\link[processx]{run}}}
}}

