% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{reload_source_files}
\alias{reload_source_files}
\alias{reload_source_files_inplace}
\title{Test if source files are current}
\usage{
reload_source_files(envir, source_files = attr(envir, "source_files"), ...)

reload_source_files_inplace(envir, ...)
}
\arguments{
\item{envir}{Environment to load}

\item{source_files}{Named character vector where names correspond
to filenames and the contents correspond to md5sums returned by
\code{link{md5sum}}.  This is the format in the attribute
\code{source_files} from \code{\link{load_source_files}} when run with
\code{collect_hash=TRUE}.}

\item{...}{Additional arguments to \code{\link{load_source_files}}.}
}
\description{
Test if environments are current and reload out-of-date
environments.
}
\details{
If source_files are reloaded, a \emph{new} environment is created;
this environment will have the same parent as \code{envir}.
Alternatively, use \code{reload_source_files_inplace} to delete
all the objects from \code{envir} and re-source into that
environment though.  Beware that that will affect anything else
that uses this environment.

Test if source_files are current, based on their md5 signature.
Nothing clever is done here; missing files, changed files, etc all
return \code{FALSE}.
}

