% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript.R
\name{rscript_process}
\alias{rscript_process}
\title{External \code{Rscript} process}
\description{
An \verb{Rscript script.R} command that runs in the background. This is an
R6 class that extends the \link{process} class.
}
\section{Usage}{
\preformatted{rp <- rscript_process$new(options)
}
}

\section{Arguments}{

\itemize{
\item \code{options} A list of options created via \code{\link[=rscript_process_options]{rscript_process_options()}}.
}
}

\section{Details}{

\code{rscript_process$new} creates a new instance. Its \code{options} argument is
best created by the \code{\link[=rscript_process_options]{rscript_process_options()}} function.
}

\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
options <- rscript_process_options(script = "script.R")
rp <- rscript_process$new(options)
rp$wait()
rp$read_output_lines()
\dontshow{\}) # examplesIf}
}
