% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rscript_process_options}
\alias{rscript_process_options}
\title{Create options for an \link{rscript_process} object}
\usage{
rscript_process_options(...)
}
\arguments{
\item{...}{Options to override, named arguments.}
}
\value{
A list of options.

\code{rscript_process_options()} creates a set of options to initialize a new
object from the \code{rscript_process} class. Its arguments must be named,
the names are used as option names. The options correspond to (some of)
the arguments of the \code{\link[=rscript]{rscript()}} function. At least the \code{script} option
must be specified, the script file to run.
}
\description{
Create options for an \link{rscript_process} object

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\examples{
## List all options and their default values:
rscript_process_options()
}
