% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/system.R
\name{Sys_which}
\alias{Sys_which}
\title{Find full path to executable}
\usage{
Sys_which(name)
}
\arguments{
\item{name}{Name or path of possible exectutable to look for}
}
\description{
Like \code{Sys.which} but throws error of command not found.
}
\examples{
try(Sys_which("R"))
try(Sys_which("not_installed_program"))
}

