% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{callr}
\alias{callr}
\title{Call R from R}
\description{
It is sometimes useful to perform a computation in a separate R
process, without affecting the current R process at all. This packages
does exactly that.
}
