% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{r_eval}
\alias{r_eval}
\title{Evaluate an expression in another R session}
\usage{
r_eval(expr, mode = c("vanilla", "copycat"))
}
\arguments{
\item{expr}{Expression to evaluate.}

\item{mode}{How to set up the other R session. See details below.}
}
\value{
Value of the evaluated expression.
}
\description{
Supported modes: \itemize{
\item \sQuote{vanilla} Do not load any startup files, and
  use a temporary empty package library. It tries to imitate
  an empty vanilla R session. Note that it does not set
  \code{options("repos")} currently, so if you want to use
  a CRAN mirror, you need to set that.
\item \sQuote{copycat} Try to mimic the settings from the
  current R session: startup files, package libraries,
  loaded and attached packages, and the \code{options("repos")}
  settings will be replicated in the new R session. Not yet
  implemented.
}
}
\examples{

# Workspace is empty in vanilla mode
r_eval(ls())

# library path is "empty" in vanilla mode
r_eval(.libPaths())

# This should ideally only include base and recommended packages
r_eval(installed.packages())

}

