% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-session.R
\name{r_session_options}
\alias{r_session_options}
\title{Create options for an \link{r_session} object}
\usage{
r_session_options(...)
}
\arguments{
\item{...}{Options to override, named arguments.}
}
\description{
Create options for an \link{r_session} object
}
