% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{rcmd_process_options}
\alias{rcmd_process_options}
\title{Create options for an \link{rcmd_process} object}
\usage{
rcmd_process_options(...)
}
\arguments{
\item{...}{Options to override, named arguments.}
}
\value{
A list of options.

\code{rcmd_process_options()} creates a set of options to initialize a new
object from the \code{rcmd_process} class. Its arguments must be named, the
names are used as option names. The options correspond to (some of)
the arguments of the \code{\link[=rcmd]{rcmd()}} function. At least the \code{cmd} option must
be specified, to select the \verb{R CMD} subcommand to run. Typically
\code{cmdargs} is specified as well, to supply more arguments to \verb{R CMD}.
}
\description{
Create options for an \link{rcmd_process} object

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\examples{
## List all options and their default values:
rcmd_process_options()
}
