% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmd.R
\name{rcmd_safe_env}
\alias{rcmd_safe_env}
\title{\code{rcmd_safe_env} returns a set of environment variables that are
more appropriate for \code{rcmd_safe}.}
\usage{
rcmd_safe_env()
}
\description{
\code{rcmd_safe_env} returns a set of environment variables that are
more appropriate for \code{rcmd_safe}.
}
