\docType{import}
\name{reexports}
\alias{reexports}
\alias{run}
\alias{process}
\alias{poll}
\title{Objects exported from other packages}
\description{
\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}

These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{processx}{\code{\link[processx]{poll}}, \code{\link[processx]{process}}, \code{\link[processx]{run}}}
}}
\keyword{internal}
